+++ local/setup.sh
@@ -3,6 +3,9 @@
 # 
 # how many threads do we have?
 threads=12
+pull_kraken=true # NB - this takes a while to run and kraken2 has a habit of failing silently
+pull_humann=false # no point in pulling the humann2 databases if we're not using them atm
+
 # https://github.com/conda/conda/issues/7980
 # these are the two default locations I've encountered
 #source /etc/profile.d/conda.sh 
@@ -62,15 +65,21 @@ rm -rf databases/taxon_databases
 # set up the kraken2 database that we'll be matching our taxons agains
 mkdir -p databases/taxon_databases
 cd databases/taxon_databases
-../../kraken2/kraken2-build --download-taxonomy --db kraken_taxon --threads $threads --use-ftp  # use ftp gets ignored if it's at the beginning (?)
-../../kraken2/kraken2-build --download-library archaea --db kraken_taxon --threads $threads --use-ftp --no-masking
-../../kraken2/kraken2-build --download-library bacteria --db kraken_taxon --threads $threads --use-ftp --no-masking
-../../kraken2/kraken2-build --download-library fungi --db kraken_taxon --threads $threads --use-ftp --no-masking
-../../kraken2/kraken2-build --build --db kraken_taxon --threads $threads
-rm -rf kraken_taxon/library # get rid of the 4 gig library source files
-# set up the refence database that we'll be using to filter the reads (note that it's the GRCh38 reference)
-mkdir human_reference
-mv kraken_taxon/taxonomy human_reference/taxonomy # this takes up around 30 gigs - if we can avoid downloading it again we should
+if [ "$pull_kraken" = true ] ; then
+    ../../kraken2/kraken2-build --download-taxonomy --db kraken_taxon --threads $threads --use-ftp  # use ftp gets ignored if it's at the beginning (?)
+    ../../kraken2/kraken2-build --download-library archaea --db kraken_taxon --threads $threads --use-ftp --no-masking
+    ../../kraken2/kraken2-build --download-library bacteria --db kraken_taxon --threads $threads --use-ftp --no-masking
+    ../../kraken2/kraken2-build --download-library fungi --db kraken_taxon --threads $threads --use-ftp --no-masking
+    ../../kraken2/kraken2-build --build --db kraken_taxon --threads $threads
+    rm -rf kraken_taxon/library # get rid of the 4 gig library source files
+    # set up the refence database that we'll be using to filter the reads (note that it's the GRCh38 reference)
+    mkdir human_reference
+    mv kraken_taxon/taxonomy human_reference/taxonomy # this takes up around 30 gigs - if we can avoid downloading it again we should
+fi
+if [ "$pull_kraken" = false ] ; then
+    # taxonomies didn't get pulled for the taxon base - need to pull them now
+    ../../kraken2/kraken2-build --download-taxonomy --db human_reference --threads $threads --use-ftp
+fi
 ../../kraken2/kraken2-build --use-ftp --no-masking --download-library human --db human_reference --threads $threads
 ../../kraken2/kraken2-build --build --db human_reference --threads $threads
 rm -rf human_reference/library # get rid of the 76 gig taxonomy library files
@@ -96,23 +105,26 @@ touch input/1_1.fq.gz
 touch input/1_2.fq.gz
 # build up the databases using stag
 snakemake create_groot_index --cores $threads
-# set up metaphlan
-metaphlan --install
-# humann2 is being needlesly annoying due to dependency conflicts
-# will just rip the commands out of stag and run it as is
-# snakemake download_humann2_databases --cores 12
-#
-# setup a conda env running python 2
-# pipe yes to overwrite the env
-echo "y" | conda create --name humann2 python=2
-conda activate humann2
-# and add needed channels
-conda config --add channels defaults
-conda config --add channels bioconda
-conda config --add channels conda-forge
-# pipe yes into the install to silence prompts
-echo "y" | conda install -c bioconda -c conda-forge humann2==2.8.1 
-# download_humann2_databases
-cd ../../.. # path out of the stag copy and move back to the base dir
-humann2_databases --download chocophlan full databases/func_databases/humann2
-humann2_databases --download uniref uniref90_diamond databases/func_databases/humann2
+
+if [ "$pull_humann" = true ] ; then
+    # set up metaphlan
+    metaphlan --install # I'm pretty sure we only need metaphlan if we're dealing with humann2
+    # humann2 is being needlesly annoying due to dependency conflicts
+    # will just rip the commands out of stag and run it as is
+    # snakemake download_humann2_databases --cores 12
+    #
+    # setup a conda env running python 2
+    # pipe yes to overwrite the env
+    echo "y" | conda create --name humann2 python=2
+    conda activate humann2
+    # and add needed channels
+    conda config --add channels defaults
+    conda config --add channels bioconda
+    conda config --add channels conda-forge
+    # pipe yes into the install to silence prompts
+    echo "y" | conda install -c bioconda -c conda-forge humann2==2.8.1 
+    # download_humann2_databases
+    cd ../../.. # path out of the stag copy and move back to the base dir
+    humann2_databases --download chocophlan full databases/func_databases/humann2
+    humann2_databases --download uniref uniref90_diamond databases/func_databases/humann2
+fi
